fix: remove invoices mocks
