--- backend/src/tmp-mocked-data/invoices.ts
@@ -1,310 +1,0 @@
-import { InvoiceOrigin, InvoicesResponse, InvoiceStatus, InvoiceType } from '@/data-contracts/invoices/data-contracts';
-
-// Used as test-data. Remove when test-invoices can be fetched from api reliably
-export const mockedInvoices: InvoicesResponse['invoices'] = [
-  {
-    dueDate: '2024-06-17',
-    totalAmount: 814,
-    amountVatIncluded: 813.5,
-    amountVatExcluded: 651.2,
-    vatEligibleAmount: 651.2,
-    rounding: 0.5,
-    vat: 162.8,
-    reversedVat: false,
-    pdfAvailable: false,
-    currency: 'SEK',
-    invoiceDate: '2024-06-17',
-    fromDate: '2024-06-17',
-    toDate: '2024-06-17',
-    invoiceNumber: '999',
-    invoiceStatus: InvoiceStatus.PAID,
-    ocrNumber: '96758235',
-    organizationNumber: 'XXXXXXXXXX',
-    invoiceName: 'faktura-999.pdf',
-    invoiceType: InvoiceType.INVOICE,
-    invoiceDescription: InvoiceStatus.PAID,
-    invoiceAddress: {
-      street: 'Storgatangrändvägen 1',
-      postcode: '31532',
-      city: 'Sundsvall',
-      careOf: 'Person',
-    },
-    facilityId: 'string',
-    invoiceOrigin: InvoiceOrigin.PUBLIC_ADMINISTRATION,
-  },
-  {
-    dueDate: '2024-09-17',
-    totalAmount: 814,
-    amountVatIncluded: 813.5,
-    amountVatExcluded: 651.2,
-    vatEligibleAmount: 651.2,
-    rounding: 0.5,
-    vat: 162.8,
-    reversedVat: false,
-    pdfAvailable: false,
-    currency: 'SEK',
-    invoiceDate: '2024-06-17',
-    fromDate: '2024-06-17',
-    toDate: '2024-09-17',
-    invoiceNumber: '999',
-    invoiceStatus: InvoiceStatus.SENT,
-    ocrNumber: '96758235',
-    organizationNumber: 'XXXXXXXXXX',
-    invoiceName: 'faktura-999.pdf',
-    invoiceType: InvoiceType.INVOICE,
-    invoiceDescription: InvoiceStatus.SENT,
-    invoiceAddress: {
-      street: 'Storgatangrändvägen 1',
-      postcode: '31532',
-      city: 'Sundsvall',
-      careOf: 'Person',
-    },
-    facilityId: 'string',
-    invoiceOrigin: InvoiceOrigin.PUBLIC_ADMINISTRATION,
-  },
-  {
-    dueDate: '2024-06-17',
-    totalAmount: 814,
-    amountVatIncluded: 813.5,
-    amountVatExcluded: 651.2,
-    vatEligibleAmount: 651.2,
-    rounding: 0.5,
-    vat: 162.8,
-    reversedVat: false,
-    pdfAvailable: false,
-    currency: 'SEK',
-    invoiceDate: '2024-06-17',
-    fromDate: '2024-06-17',
-    toDate: '2024-06-17',
-    invoiceNumber: '999',
-    invoiceStatus: InvoiceStatus.PARTIALLY_PAID,
-    ocrNumber: '96758235',
-    organizationNumber: 'XXXXXXXXXX',
-    invoiceName: 'faktura-999.pdf',
-    invoiceType: InvoiceType.INVOICE,
-    invoiceDescription: InvoiceStatus.PARTIALLY_PAID,
-    invoiceAddress: {
-      street: 'Storgatangrändvägen 1',
-      postcode: '31532',
-      city: 'Sundsvall',
-      careOf: 'Person',
-    },
-    facilityId: 'string',
-    invoiceOrigin: InvoiceOrigin.PUBLIC_ADMINISTRATION,
-  },
-  {
-    dueDate: '2024-03-17',
-    totalAmount: 814,
-    amountVatIncluded: 813.5,
-    amountVatExcluded: 651.2,
-    vatEligibleAmount: 651.2,
-    rounding: 0.5,
-    vat: 162.8,
-    reversedVat: false,
-    pdfAvailable: false,
-    currency: 'SEK',
-    invoiceDate: '2024-01-17',
-    fromDate: '2024-01-17',
-    toDate: '2024-03-17',
-    invoiceNumber: '999',
-    invoiceStatus: InvoiceStatus.DEBT_COLLECTION,
-    ocrNumber: '96758235',
-    organizationNumber: 'XXXXXXXXXX',
-    invoiceName: 'faktura-999.pdf',
-    invoiceType: InvoiceType.INVOICE,
-    invoiceDescription: InvoiceStatus.DEBT_COLLECTION,
-    invoiceAddress: {
-      street: 'Storgatangrändvägen 1',
-      postcode: '31532',
-      city: 'Sundsvall',
-      careOf: 'Person',
-    },
-    facilityId: 'string',
-    invoiceOrigin: InvoiceOrigin.PUBLIC_ADMINISTRATION,
-  },
-  {
-    dueDate: '2024-06-17',
-    totalAmount: 814,
-    amountVatIncluded: 813.5,
-    amountVatExcluded: 651.2,
-    vatEligibleAmount: 651.2,
-    rounding: 0.5,
-    vat: 162.8,
-    reversedVat: false,
-    pdfAvailable: false,
-    currency: 'SEK',
-    invoiceDate: '2024-06-17',
-    fromDate: '2024-06-17',
-    toDate: '2024-06-17',
-    invoiceNumber: '999',
-    invoiceStatus: InvoiceStatus.PAID_TOO_MUCH,
-    ocrNumber: '96758235',
-    organizationNumber: 'XXXXXXXXXX',
-    invoiceName: 'faktura-999.pdf',
-    invoiceType: InvoiceType.INVOICE,
-    invoiceDescription: InvoiceStatus.PAID_TOO_MUCH,
-    invoiceAddress: {
-      street: 'Storgatangrändvägen 1',
-      postcode: '31532',
-      city: 'Sundsvall',
-      careOf: 'Person',
-    },
-    facilityId: 'string',
-    invoiceOrigin: InvoiceOrigin.PUBLIC_ADMINISTRATION,
-  },
-  {
-    dueDate: '2024-06-17',
-    totalAmount: 814,
-    amountVatIncluded: 813.5,
-    amountVatExcluded: 651.2,
-    vatEligibleAmount: 651.2,
-    rounding: 0.5,
-    vat: 162.8,
-    reversedVat: false,
-    pdfAvailable: false,
-    currency: 'SEK',
-    invoiceDate: '2024-06-17',
-    fromDate: '2024-06-17',
-    toDate: '2024-06-17',
-    invoiceNumber: '999',
-    invoiceStatus: InvoiceStatus.REMINDER,
-    ocrNumber: '96758235',
-    organizationNumber: 'XXXXXXXXXX',
-    invoiceName: 'faktura-999.pdf',
-    invoiceType: InvoiceType.INVOICE,
-    invoiceDescription: InvoiceStatus.REMINDER,
-    invoiceAddress: {
-      street: 'Storgatangrändvägen 1',
-      postcode: '31532',
-      city: 'Sundsvall',
-      careOf: 'Person',
-    },
-    facilityId: 'string',
-    invoiceOrigin: InvoiceOrigin.PUBLIC_ADMINISTRATION,
-  },
-  {
-    dueDate: '2024-06-17',
-    totalAmount: 814,
-    amountVatIncluded: 813.5,
-    amountVatExcluded: 651.2,
-    vatEligibleAmount: 651.2,
-    rounding: 0.5,
-    vat: 162.8,
-    reversedVat: false,
-    pdfAvailable: false,
-    currency: 'SEK',
-    invoiceDate: '2024-06-17',
-    fromDate: '2024-06-17',
-    toDate: '2024-06-17',
-    invoiceNumber: '999',
-    invoiceStatus: InvoiceStatus.VOID,
-    ocrNumber: '96758235',
-    organizationNumber: 'XXXXXXXXXX',
-    invoiceName: 'faktura-999.pdf',
-    invoiceType: InvoiceType.INVOICE,
-    invoiceDescription: InvoiceStatus.VOID,
-    invoiceAddress: {
-      street: 'Storgatangrändvägen 1',
-      postcode: '31532',
-      city: 'Sundsvall',
-      careOf: 'Person',
-    },
-    facilityId: 'string',
-    invoiceOrigin: InvoiceOrigin.PUBLIC_ADMINISTRATION,
-  },
-  {
-    dueDate: '2024-06-17',
-    totalAmount: 814,
-    amountVatIncluded: 813.5,
-    amountVatExcluded: 651.2,
-    vatEligibleAmount: 651.2,
-    rounding: 0.5,
-    vat: 162.8,
-    reversedVat: false,
-    pdfAvailable: false,
-    currency: 'SEK',
-    invoiceDate: '2024-06-17',
-    fromDate: '2024-06-17',
-    toDate: '2024-06-17',
-    invoiceNumber: '999',
-    invoiceStatus: InvoiceStatus.CREDITED,
-    ocrNumber: '96758235',
-    organizationNumber: 'XXXXXXXXXX',
-    invoiceName: 'faktura-999.pdf',
-    invoiceType: InvoiceType.INVOICE,
-    invoiceDescription: InvoiceStatus.CREDITED,
-    invoiceAddress: {
-      street: 'Storgatangrändvägen 1',
-      postcode: '31532',
-      city: 'Sundsvall',
-      careOf: 'Person',
-    },
-    facilityId: 'string',
-    invoiceOrigin: InvoiceOrigin.PUBLIC_ADMINISTRATION,
-  },
-  {
-    dueDate: '2024-06-17',
-    totalAmount: 814,
-    amountVatIncluded: 813.5,
-    amountVatExcluded: 651.2,
-    vatEligibleAmount: 651.2,
-    rounding: 0.5,
-    vat: 162.8,
-    reversedVat: false,
-    pdfAvailable: false,
-    currency: 'SEK',
-    invoiceDate: '2024-06-17',
-    fromDate: '2024-06-17',
-    toDate: '2024-06-17',
-    invoiceNumber: '999',
-    invoiceStatus: InvoiceStatus.WRITTEN_OFF,
-    ocrNumber: '96758235',
-    organizationNumber: 'XXXXXXXXXX',
-    invoiceName: 'faktura-999.pdf',
-    invoiceType: InvoiceType.INVOICE,
-    invoiceDescription: InvoiceStatus.WRITTEN_OFF,
-    invoiceAddress: {
-      street: 'Storgatangrändvägen 1',
-      postcode: '31532',
-      city: 'Sundsvall',
-      careOf: 'Person',
-    },
-    facilityId: 'string',
-    invoiceOrigin: InvoiceOrigin.PUBLIC_ADMINISTRATION,
-  },
-  {
-    dueDate: '2024-06-17',
-    totalAmount: 814,
-    amountVatIncluded: 813.5,
-    amountVatExcluded: 651.2,
-    vatEligibleAmount: 651.2,
-    rounding: 0.5,
-    vat: 162.8,
-    reversedVat: false,
-    pdfAvailable: false,
-    currency: 'SEK',
-    invoiceDate: '2024-06-17',
-    fromDate: '2024-06-17',
-    toDate: '2024-06-17',
-    invoiceNumber: '999',
-    invoiceStatus: InvoiceStatus.UNKNOWN,
-    ocrNumber: '96758235',
-    organizationNumber: 'XXXXXXXXXX',
-    invoiceName: 'faktura-999.pdf',
-    invoiceType: InvoiceType.INVOICE,
-    invoiceDescription: InvoiceStatus.UNKNOWN,
-    invoiceAddress: {
-      street: 'Storgatangrändvägen 1',
-      postcode: '31532',
-      city: 'Sundsvall',
-      careOf: 'Person',
-    },
-    facilityId: 'string',
-    invoiceOrigin: InvoiceOrigin.PUBLIC_ADMINISTRATION,
-  },
-];
-
-export const mockedInvoiceResponse: InvoicesResponse = {
-  invoices: mockedInvoices,
-  _meta: undefined,
-};
